CNDB-17102: Break out VectorCompressionTest to reduce probability of timeouts (#2273)
### What is the issue

Fixes: https://github.com/riptano/cndb/issues/17102

### What does this PR fix and why was it fixed

This PR breaks the `VectorCompressionTest` into a test class per dataset
to prevent the likelihood of timeouts.
