+++ test/unit/org/apache/cassandra/index/sai/disk/vector/vectorcompression/VectorCompressionGeckoTest.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.cassandra.index.sai.disk.vector.vectorcompression;
+
+import java.io.IOException;
+
+import org.junit.Test;
+
+import org.apache.cassandra.index.sai.disk.vector.VectorSourceModel;
+
+public class VectorCompressionGeckoTest extends AbstractVectorCompressionTest
+{
+    @Test
+    public void testGecko() throws IOException
+    {
+        // GECKO is always 768
+        testOne(VectorSourceModel.GECKO, 768, VectorSourceModel.GECKO.compressionProvider.apply(768));
+    }
+}
